Fixed bugs

+++ src/fusion/distance/Distance.java
@@ -3,5 +3,5 @@ package fusion.distance;
 import java.util.Collection;
 //Abstract Distance function
 public abstract class Distance {
-	public abstract int computeDistance(Collection<String> Ext, Collection<String> Mod);
+	public abstract float computeDistance(Collection<String> Ext, Collection<String> Mod);
 }
