+++ src/fusion/distance/DistanceHamming.java
@@ -4,8 +4,8 @@ import java.util.Collection;
 public class DistanceHamming extends Distance {
 
     // Function to calculate the hamming distance 
-    public int computeDistance(Collection<String> Ext, Collection<String> Mod) {
-        int difference = 0;
+    public float computeDistance(Collection<String> Ext, Collection<String> Mod) {
+        float difference = 0f;
         for (String i : Ext) {
             if (!Mod.contains(i)) {
                 difference++;
